divide h and cpp
