+++ src/list.h
@@ -14,50 +14,16 @@ enum GPtype {
 };
 
 struct List {
-    ~List() {
-        clear();
-    }
+    ~List();
 
-    void init(int n) {
-        clear();
-        vals = new void* [n];
-        names = new char* [n];
-        forms = new char* [n];
-        types = new GPtype [n];
-        skips = new bool [n];
-        am = n;
-    }
-    
-    void add(void* data, const char* name, GPtype type) {
-        if (idx < am) {
-            vals[idx] = data;
-            names[idx] = (char*)name;
-            types[idx] = type;
-            skips[idx] = true;
-            idx++;
-        }
-    }
-    
-    void add(void* data, const char* form, const char* name, GPtype type) {
-        if (idx < am) {
-            vals[idx] = data;
-            forms[idx] = (char*)form;
-            names[idx] = (char*)name;
-            types[idx] = type;
-            skips[idx] = false;
-            idx++;
-        }
-    }
+    void init(int n);
 
-    void clear() {
-        if (vals) delete [] vals;
-        if (names) delete [] names;
-        if (forms) delete [] forms;
-        if (types) delete [] types;
-        if (skips) delete [] skips;
-        idx = am = 0;
-    }
-    
+    void add(void* data, const char* name, GPtype type);
+
+    void add(void* data, const char* form, const char* name, GPtype type);
+
+    void clear();
+
     int idx = 0;
     int am = 0;
     void** vals = nullptr;
